requirements.txt
appdirs==1.4.4
asgiref==3.2.7
attrs==19.3.0
coverage==5.1
distlib==0.3.0
Django==3.0.6
filelock==3.0.12
image==1.5.32
importlib-metadata==1.6.0
more-itertools==8.3.0
mypy==0.770
mypy-extensions==0.4.3
NuMPI==0.1.1
numpy==1.18.4
packaging==20.4
Pillow==7.1.2
pluggy==0.13.1
py==1.8.1
pyparsing==2.4.7
pytest==5.4.2
pytz==2020.1
RPi.GPIO==0.7.0
runtests==0.0.28
spidev==3.4
sqlparse==0.3.1
typed-ast==1.4.1
typing-extensions==3.7.4.2
virtualenv==20.0.21
wcwidth==0.1.9
zipp==3.1.0
